requirements.txt
# The official Model Context Protocol SDK
mcp[cli]>=1.0.0

# Web Framework and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data Validation (Core to MCP & FastAPI)
pydantic>=2.0.0

# Configuration management
python-dotenv>=1.0.0

openai>=1.0.0
streamlit>=1.30.0